Kartaca-Task-Content-Access/requirements.txt
websocket-client~=1.5.1
websockets~=10.4